pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "photosorter"
version = "0.1.0"
description = "Reorder travel photos by visual similarity using DINOv3 embeddings"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=1.0.20",
    "Pillow>=9.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "tqdm>=4.60",
    "rawpy>=0.20",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
photosorter = "photosorter.main:main"
photosorter-json = "photosorter_bridge.cli_json:main"

[tool.setuptools.packages.find]
where = ["engine/photosorter_core", "engine/photosorter_bridge"]
include = ["photosorter", "photosorter.*", "photosorter_bridge", "photosorter_bridge.*"]

[tool.pytest.ini_options]
testpaths = ["tests/python"]
pythonpath = ["engine/photosorter_core", "engine/photosorter_bridge"]

[tool.coverage.run]
source = ["photosorter", "photosorter_bridge"]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true
ignore_errors = true
